refactor: implement buffer-based TDOA measurement chain
The measurement chain is now:
1. Sync event at sub-sample position within a buffer (~1.6 us precision)
2. Exact integer buffer count from sync to detection point
3. Detection point is a search hint, not a timing reference
4. Server xcorr finds the PA knee position within the full snippet
5. Total TDOA = (sync_to_detection + xcorr_knee_offset) per node

Key changes:

Node:
- sample_index back to int (detection point at integer window boundary)
- Pipeline mapping back to integer division (exact for integer positions)
- The node does NOT find the knee — it sends the raw snippet and the
  server uses xcorr

Server:
- Remove ALL snippet trimming from cross_correlate_snippets: no
  transition bounds, no event_type-based fixed-fraction trimming
- Full-snippet xcorr sees the PA transition at different absolute
  positions in the two nodes' snippets (because detection points
  differ), measuring the offset that the coarse sync_delta needs
- xcorr applied before disambiguation (previous commit)

The previous trimming (both transition-bound and event_type) centered
the PA feature at the same relative position in both snippets,
producing ~0 xcorr lag and defeating the detection-point correction.

+++ src/beagle_node/pipeline/carrier_detect.py
@@ -65,22 +65,22 @@ _TIMING_DIAG = os.environ.get("BEAGLE_TIMING_DIAG") == "1"
 @dataclass(frozen=True)
 class CarrierOnset:
     """LMR carrier has appeared."""
-    sample_index: float         # Sub-sample knee position in the continuous IQ stream
+    sample_index: int           # Detection point in the continuous IQ stream (window boundary)
     power_db: float             # Instantaneous power at detection
     noise_floor_db: float = -100.0  # EMA of idle-state power; used to compute SNR
     iq_snippet: bytes = b""    # int8-interleaved IQ bytes for cross-correlation
-    transition_start: int = 0  # Sample index within snippet where transition begins
-    transition_end: int = 0    # Sample index within snippet where transition ends (plateau)
+    transition_start: int = 0  # Approximate transition zone start within snippet
+    transition_end: int = 0    # Approximate transition zone end within snippet
 
 
 @dataclass(frozen=True)
 class CarrierOffset:
     """LMR carrier has disappeared."""
-    sample_index: float         # Sub-sample knee position in the continuous IQ stream
+    sample_index: int           # Detection point in the continuous IQ stream (window boundary)
     power_db: float
     iq_snippet: bytes = b""    # int8-interleaved IQ bytes for cross-correlation
-    transition_start: int = 0  # Sample index within snippet where transition begins (plateau edge)
-    transition_end: int = 0    # Sample index within snippet where transition ends (noise)
+    transition_start: int = 0  # Approximate transition zone start within snippet
+    transition_end: int = 0    # Approximate transition zone end within snippet
 
 
 _State = Literal["idle", "active"]
